+++ LeftShoeEmporium/src/com/cognixia/jump/dao/ShoeDAO.java
@@ -37,7 +37,7 @@ public class ShoeDAO {
 				shoe.setShoe_name(rs.getString(3));
 				shoe.setShoe_type(rs.getString(4));
 				shoe.setShoe_code(rs.getString(5));
-				shoe.setPrice(rs.getFloat(6));
+				shoe.setPrice(rs.getDouble(6));
 				shoe.setStock(rs.getInt(7));
 
 				shoes.add(shoe);
@@ -83,7 +83,7 @@ public class ShoeDAO {
 				shoe.setShoe_name(rs.getString(3));
 				shoe.setShoe_type(rs.getString(4));
 				shoe.setShoe_code(rs.getString(5));
-				shoe.setPrice(rs.getFloat(6));
+				shoe.setPrice(rs.getDouble(6));
 				shoe.setStock(rs.getInt(7));
 
 			}
